requirements.txt
requests==2.28.2
pycryptodome==3.16.0
pyinstaller==5.7.0